Improved C long handling for C#
Defining SWIGWORDSIZE64 now applies the (unsigned)
long long typemaps to (unsigned) long for a better match on systems
where long is 64-bits. A new "Type mapping" section has been added into
the CSharp.html documentation covering this and marshalling of primitive
types. C (unsigned) long handling remains as is by default, that is,
marshall as 32-bit.

The INPUT[], OUTPUT[], INOUT[], FIXED[] typemaps for long and unsigned long
in arrays_csharp.i can now be used and compiled on 64-bit platforms where
sizeof(long) != sizeof(int). Requires SWIGWORDSIZE64 to be defined.

Move SWIGWORDSIZE64 check into fragments

Changes also made so that the C# test-suite passes using:
  env SWIG_FEATURES=-DSWIGWORDSIZE64 make check-csharp-test-suite

See issue #2379

+++ Examples/test-suite/csharp/li_boost_shared_ptr_runme.cs
@@ -591,7 +591,7 @@ public class runme
       throw new Exception("verify value failed. Expected: " + expected + " Got: " + got);
   }
   private void verifyCount(int expected, Klass k) {
-    int got = li_boost_shared_ptr.use_count(k); 
+    int got = (int)li_boost_shared_ptr.use_count(k); 
     if (expected != got)
       throw new Exception("verify use_count failed. Expected: " + expected + " Got: " + got);
   }
